Reapply "Merge branch 'multi-floor-feature' into lua-steam"
This reverts commit a8dd4f56214768d2e681b25c7e3fbf0eb9e10a3f.
