+++ src/main/java/cam72cam/immersiverailroading/entity/EntityMoveableRollingStock.java
@@ -30,7 +30,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public abstract class EntityMoveableRollingStock extends EntityCustomPlayerMovement implements ICollision {
+public abstract class EntityMoveableRollingStock extends EntityRidableRollingStock implements ICollision {
     @TagField("frontYaw")
     private Float frontYaw;
     @TagField("rearYaw")
